Added Desktop launcher
Quicker for faster running than emulators or old phones. Avoid resizing
the window during runtime though, the code isn't built for it, and it
wont correctly read mouse input if you do.  I amn't sure if the file for
the desktop run configuration is included, but if not just ask me and I
can put it up on facebook. Also tried to fix an apparent memory leak,
but that may have just been my phone showing its age.