Another few minor changes to UI
Implemented the "broken" function to draw sprites for enemies and
experimented with the placement of the Health bars for enemies. Didn't
remove "Master Debater" or "Mast. Debater", but will do so with the next
commit. There's a reason why they don't call Expert Debaters a master. Its
also too long a word, I want to implement a restriction on string size so
that I can actually fit them on the screen. Cause, ya know, people have to
be able to see emm Theyre small enough as it is, and everything it
starting to become very cramped. MAX SIZE IS NOW 10 CHARACTERS.