Changed Information that the Win State provides
I fixed the printout for the rating system. It was incrementing the counter in the draw function rather than just after the combo, so it would increment multiple times per combo. At some point I will add the combo score into the rating system somehow, but it will be a lst minute kind of thing. There is now also a second page to the information, but it is not fully implemented yet. This functionality will be added in the next commit later.