Swapped how rotation works to fit all devices
It originally rotated at an angle defined by fps for a second. This
means that any change in fps during rotation would through it off. Now
it rotates by a set amount each frame, for a time longer than .8
seconds, until it rotates back to zero. It ends at the same point it
starts on every device now. Also finally gave Recharge a function, not
sure how well balanced it is though