2 menu types, enemy damage item fixed
the fight menu type is now toggled by boolean menuStyle2, false gives
the old menu style, true will give the new style with 4 icons per page
all of which can be tapped. The drawing for this is currently working
but tapping isn't done yet so the default is the old style for now.
Small changes to the new enemy damage item so it works, and so its
quantity will now be displayed in inventory.