Square combat menu working
toggles in the game menu are working as well, default is targeting
interface off and square menu on. Just need to change the font size in
the square menu and uncomment out the tooltip code. Small change so
there's always a back button whether the targeting arrows are on or not,
so you can get out of targeting.